+++ ErsatzTV.Core/Scheduling/YamlScheduling/Models/YamlPostRollInstruction.cs
@@ -1,0 +1,11 @@
+using YamlDotNet.Serialization;
+
+namespace ErsatzTV.Core.Scheduling.YamlScheduling.Models;
+
+public class YamlPostRollInstruction : YamlPlayoutInstruction
+{
+    [YamlMember(Alias = "post_roll", ApplyNamingConventions = false)]
+    public bool PostRoll { get; set; }
+
+    public string Sequence { get; set; }
+}