+++ CHANGELOG.md
@@ -15,6 +15,9 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 - YAML playout: add `pre_roll` instruction to enable and disable a pre-roll sequence
   - With value of `true` and `sequence` property, will enable automatic pre-roll for all content in the playout to the sequence with the provided key
   - With value of `false`, will disable automatic pre-roll in the playout
+- YAML playout: add `post_roll` instruction to enable and disable a post-roll sequence
+    - With value of `true` and `sequence` property, will enable automatic post-roll for all content in the playout to the sequence with the provided key
+    - With value of `false`, will disable automatic post-roll in the playout
 - Add YAML playout validation (using JSON Schema)
   - Invalid YAML playout definitions will fail to build and will log validation failures as warnings
   - `content` is fully validated